+++ sdk/resourcemanager/storage/armstorage/models.go
@@ -182,6 +182,46 @@ type AccountMicrosoftEndpoints struct {
 	Web *string `json:"web,omitempty" azure:"ro"`
 }
 
+// AccountMigration - The object encapsulating storage account migration properties.
+type AccountMigration struct {
+	// The parameters or status associated with an ongoing or enqueued storage account migration in order to update its current
+	// SKU or region.
+	Properties *AccountMigrationProperties `json:"properties,omitempty"`
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; The name of the resource
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; Metadata pertaining to creation and last modification of the resource.
+	SystemData *SystemData `json:"systemData,omitempty" azure:"ro"`
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
+// AccountMigrationProperties - The parameters or status associated with an ongoing or enqueued storage account migration
+// in order to update its current SKU or region.
+type AccountMigrationProperties struct {
+	// REQUIRED; The target sku name to migrate.
+	TargetSKUName *TargetSKUName `json:"targetSkuName,omitempty"`
+
+	// READ-ONLY; Current status of migration
+	MigrationStatus *MigrationStatus `json:"migrationStatus,omitempty" azure:"ro"`
+}
+
+// AccountMigrationsClientBeginPutOptions contains the optional parameters for the AccountMigrationsClient.BeginPut method.
+type AccountMigrationsClientBeginPutOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// AccountMigrationsClientGetOptions contains the optional parameters for the AccountMigrationsClient.Get method.
+type AccountMigrationsClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
 // AccountProperties - Properties of the storage account.
 type AccountProperties struct {
 	// Allow or disallow public access to all blobs or containers in the storage account. The default interpretation is true for
@@ -250,6 +290,9 @@ type AccountProperties struct {
 	// be changed for the premium block blobs storage account type.
 	AccessTier *AccessTier `json:"accessTier,omitempty" azure:"ro"`
 
+	// READ-ONLY; If customer initiated account migration is in progress, the value will be true else it will be null.
+	AccountMigrationInProgress *bool `json:"accountMigrationInProgress,omitempty" azure:"ro"`
+
 	// READ-ONLY; Blob restore status
 	BlobRestoreStatus *BlobRestoreStatus `json:"blobRestoreStatus,omitempty" azure:"ro"`
 
@@ -267,6 +310,9 @@ type AccountProperties struct {
 
 	// READ-ONLY; Geo Replication Stats
 	GeoReplicationStats *GeoReplicationStats `json:"geoReplicationStats,omitempty" azure:"ro"`
+
+	// READ-ONLY; This property value will be set as true or false as migration progresses. Default value is null.
+	IsSKUConversionBlocked *bool `json:"isSkuConversionBlocked,omitempty" azure:"ro"`
 
 	// READ-ONLY; Storage account keys creation time.
 	KeyCreationTime *KeyCreationTime `json:"keyCreationTime,omitempty" azure:"ro"`
@@ -562,6 +608,9 @@ type AccountsClientBeginCreateOptions struct {
 
 // AccountsClientBeginFailoverOptions contains the optional parameters for the AccountsClient.BeginFailover method.
 type AccountsClientBeginFailoverOptions struct {
+	// The parameter is set to 'Planned' to indicate whether a Planned failover is requested.. Specifying any value will set the
+	// value to Planned.
+	FailoverType *string
 	// Resumes the LRO from the provided token.
 	ResumeToken string
 }
@@ -1085,27 +1134,6 @@ type CheckNameAvailabilityResult struct {
 	Reason *Reason `json:"reason,omitempty" azure:"ro"`
 }
 
-// CloudError - An error response from the Storage service.
-type CloudError struct {
-	// An error response from the Storage service.
-	Error *CloudErrorBody `json:"error,omitempty"`
-}
-
-// CloudErrorBody - An error response from the Storage service.
-type CloudErrorBody struct {
-	// An identifier for the error. Codes are invariant and are intended to be consumed programmatically.
-	Code *string `json:"code,omitempty"`
-
-	// A list of additional details about the error.
-	Details []*CloudErrorBody `json:"details,omitempty"`
-
-	// A message describing the error, intended to be suitable for display in a user interface.
-	Message *string `json:"message,omitempty"`
-
-	// The target of the particular error. For example, the name of the property in error.
-	Target *string `json:"target,omitempty"`
-}
-
 // ContainerProperties - The properties of a container.
 type ContainerProperties struct {
 	// Default the container to use specified encryption scope for all writes.
@@ -1424,7 +1452,12 @@ type EncryptionScopesClientGetOptions struct {
 
 // EncryptionScopesClientListOptions contains the optional parameters for the EncryptionScopesClient.List method.
 type EncryptionScopesClientListOptions struct {
-	// placeholder for future optional parameters
+	// Optional. When specified, only encryption scope names starting with the filter will be listed.
+	Filter *string
+	// Optional, when specified, will list encryption scopes with the specific state. Defaults to All
+	Include *ListEncryptionScopesInclude
+	// Optional, specifies the maximum number of encryption scopes that will be included in the list response.
+	Maxpagesize *string
 }
 
 // EncryptionScopesClientPatchOptions contains the optional parameters for the EncryptionScopesClient.Patch method.
@@ -2236,11 +2269,17 @@ type ManagementPolicyBaseBlob struct {
 	// This property enables auto tiering of a blob from cool to hot on a blob access. This property requires tierToCool.daysAfterLastAccessTimeGreaterThan.
 	EnableAutoTierToHotFromCool *bool `json:"enableAutoTierToHotFromCool,omitempty"`
 
-	// The function to tier blobs to archive storage. Support blobs currently at Hot or Cool tier
+	// The function to tier blobs to archive storage.
 	TierToArchive *DateAfterModification `json:"tierToArchive,omitempty"`
 
-	// The function to tier blobs to cool storage. Support blobs currently at Hot tier
+	// The function to tier blobs to cold storage.
+	TierToCold *DateAfterModification `json:"tierToCold,omitempty"`
+
+	// The function to tier blobs to cool storage.
 	TierToCool *DateAfterModification `json:"tierToCool,omitempty"`
+
+	// The function to tier blobs to hot storage. This action can only be used with Premium Block Blob Storage Accounts
+	TierToHot *DateAfterModification `json:"tierToHot,omitempty"`
 }
 
 // ManagementPolicyDefinition - An object that defines the Lifecycle rule. Each definition is made up with a filters set and
@@ -2303,11 +2342,17 @@ type ManagementPolicySnapShot struct {
 	// The function to delete the blob snapshot
 	Delete *DateAfterCreation `json:"delete,omitempty"`
 
-	// The function to tier blob snapshot to archive storage. Support blob snapshot currently at Hot or Cool tier
+	// The function to tier blob snapshot to archive storage.
 	TierToArchive *DateAfterCreation `json:"tierToArchive,omitempty"`
 
-	// The function to tier blob snapshot to cool storage. Support blob snapshot currently at Hot tier
+	// The function to tier blobs to cold storage.
+	TierToCold *DateAfterCreation `json:"tierToCold,omitempty"`
+
+	// The function to tier blob snapshot to cool storage.
 	TierToCool *DateAfterCreation `json:"tierToCool,omitempty"`
+
+	// The function to tier blobs to hot storage. This action can only be used with Premium Block Blob Storage Accounts
+	TierToHot *DateAfterCreation `json:"tierToHot,omitempty"`
 }
 
 // ManagementPolicyVersion - Management policy action for blob version.
@@ -2315,11 +2360,17 @@ type ManagementPolicyVersion struct {
 	// The function to delete the blob version
 	Delete *DateAfterCreation `json:"delete,omitempty"`
 
-	// The function to tier blob version to archive storage. Support blob version currently at Hot or Cool tier
+	// The function to tier blob version to archive storage.
 	TierToArchive *DateAfterCreation `json:"tierToArchive,omitempty"`
 
-	// The function to tier blob version to cool storage. Support blob version currently at Hot tier
+	// The function to tier blobs to cold storage.
+	TierToCold *DateAfterCreation `json:"tierToCold,omitempty"`
+
+	// The function to tier blob version to cool storage.
 	TierToCool *DateAfterCreation `json:"tierToCool,omitempty"`
+
+	// The function to tier blobs to hot storage. This action can only be used with Premium Block Blob Storage Accounts
+	TierToHot *DateAfterCreation `json:"tierToHot,omitempty"`
 }
 
 // MetricSpecification - Metric specification of operation.